requirements.txt
click
colorama
geopy
ipgetter
python-dateutil
python-geoip
python-geoip-geolite2
requests
typing
win-inet-pton